factor out directrunner userstate classes
