+++ sdks/python/apache_beam/transforms/userstate_test.py
@@ -308,80 +308,5 @@ class InterfaceTest(unittest.TestCase):
       UserStateUtils.validate_stateful_dofn(dofn)
 
 
-class StatefulDoFnOnDirectRunnerTest(unittest.TestCase):
-  all_records = None
-
-  def setUp(self):
-    # Use state on the TestCase class, since other references would be pickled
-    # into a closure and not have the desired side effects.
-    StatefulDoFnOnDirectRunnerTest.all_records = []
-
-  def record_dofn(self):
-    class RecordDoFn(DoFn):
-      def process(self, element):
-        StatefulDoFnOnDirectRunnerTest.all_records.append(element)
-
-    return RecordDoFn()
-
-  def test_simple_stateful_dofn(self):
-    class SimpleTestStatefulDoFn(DoFn):
-      BUFFER_STATE = BagStateSpec('buffer', BytesCoder())
-      EXPIRY_TIMER = TimerSpec('expiry1', TimeDomain.WATERMARK)
-
-      def process(self, element, buffer=DoFn.StateParam(BUFFER_STATE),
-                  timer1=DoFn.TimerParam(EXPIRY_TIMER)):
-        buffer.add('A' + str(element))
-        timer1.set(20)
-
-      @on_timer(EXPIRY_TIMER)
-      def expiry_callback(self, buffer=DoFn.StateParam(BUFFER_STATE),
-                          timer=DoFn.TimerParam(EXPIRY_TIMER)):
-        yield ''.join(sorted(buffer.read()))
-
-    with TestPipeline() as p:
-      test_stream = (TestStream()
-                     .advance_watermark_to(10)
-                     .add_elements([1, 2])
-                     .add_elements([3])
-                     .advance_watermark_to(25)
-                     .add_elements([4]))
-      pcoll = p | test_stream
-      records = pcoll | beam.ParDo(SimpleTestStatefulDoFn())
-      records | beam.ParDo(self.record_dofn())
-
-    self.assertEqual(['A1A2A3', 'A1A2A3A4'],
-        StatefulDoFnOnDirectRunnerTest.all_records)
-
-  # def test_simple_stateful_dofn_combining(self):
-  #   class SimpleTestStatefulDoFn(DoFn):
-  #     BUFFER_STATE = CombiningValueStateSpec('buffer', IterableCoder(VarIntCoder()), ToListCombineFn())
-  #     EXPIRY_TIMER = TimerSpec('expiry1', TimeDomain.WATERMARK)
-
-  #     def process(self, element, buffer=DoFn.StateParam(BUFFER_STATE),
-  #                 timer1=DoFn.TimerParam(EXPIRY_TIMER)):
-  #       buffer.add(element)
-  #       timer1.set(20)
-
-  #     @on_timer(EXPIRY_TIMER)
-  #     def expiry_callback(self, buffer=DoFn.StateParam(BUFFER_STATE),
-  #                         timer=DoFn.TimerParam(EXPIRY_TIMER)):
-  #       print 'YO', buffer.read()
-  #       yield ''.join(str(x) for x in sorted(buffer.read()))
-
-  #   with TestPipeline() as p:
-  #     test_stream = (TestStream()
-  #                    .advance_watermark_to(10)
-  #                    .add_elements([1, 2])
-  #                    .add_elements([3])
-  #                    .advance_watermark_to(25)
-  #                    .add_elements([4]))
-  #     pcoll = p | test_stream
-  #     records = pcoll | beam.ParDo(SimpleTestStatefulDoFn())
-  #     records | beam.ParDo(self.record_dofn())
-
-  #   self.assertEqual(['123', '1234'],
-  #       StatefulDoFnOnDirectRunnerTest.all_records)
-
-
 if __name__ == '__main__':
   unittest.main()
